aad-with-angular/AuthAngular/ClientApp/src/app/fetch-data/weather-forecast.service.ts
import { Inject, Injectable } from "@angular/core";
import { HttpClient } from "@angular/common/http";
import { map } from "rxjs/operators";
import { Observable } from "rxjs";

@Injectable({
  providedIn: "root",
})
export class WeatherForecastService {
  constructor(
    private http: HttpClient,
    @Inject("BASE_URL") private baseUrl: string
  ) {}

  getForecasts(): Observable<WeatherForecast[]> {
    return this.http
      .get<WeatherForecast[]>(this.baseUrl + "weatherforecast")
      .pipe(map((res) => res));
  }
}

export interface WeatherForecast {
  date: string;
  temperatureC: number;
  temperatureF: number;
  summary: string;
}
